fix(ui): keep attachment types outside storage runtimes
